Clean up client deserialization generation logic
